fix finance

+++ website/src/pages/Finance/Invoice/index.js
@@ -133,7 +133,7 @@ export default function Invoice(props) {
         values['billing_period'] = values['billing_period'][0].format('YYYY-MM-DD') + " - " + values['billing_period'][1].format('YYYY-MM-DD');
         values['sessions'] = JSON.stringify(location.ids);
         const valueStr = total_amount.replace('€', '');
-        values['total_amount'] = parseInt(valueStr);
+        values['total_amount'] = parseFloat(valueStr);
 
         _service({
             method: 'PUT',
